Python 3 port: integer division
In Python 2 division of integers rounds down where as in Python 3
division of integers produces floats [1].  This leads to floats, instead
of expected integers, being passed to pygame APIs which generated
warnings with the latest Python 3.

First encounted warning:
    AsteroidsInfinity.py:197: DeprecationWarning: an integer is required (got type float).  Implicit conversion to integers using __int__ is deprecated, and may be removed in a future version of Python.
      self.rect = pygame.Rect(self.pos[0] - width / 2, self.pos[1], width, height)

To additionally get an exception with a backtrace, run as
"python3 -W error AsteroidsInfinity.py" to generate errors instead of
warnings.  First such encountered exception:
    Traceback (most recent call last):
      File "AsteroidsInfinity.py", line 1013, in <module>
        main()
      File "AsteroidsInfinity.py", line 627, in main
        level_text = Text("Level N/A", font, (screensize[0] / 2, 5), align = "center")
      File "AsteroidsInfinity.py", line 186, in __init__
        self.change(text)
      File "AsteroidsInfinity.py", line 198, in change
        self.rect = pygame.Rect(self.pos[0] - width / 2, self.pos[1], width, height)
    TypeError: Argument must be rect style object

Specify integer division (//) where necessary as this is compatible with
Python 2 and 3.

[1] http://python-future.org/compatible_idioms.html#division